Create Cost–Benefit Analysis of Bus Fare Subsidies under Financial Constraints: The Case of Asunción, Paraguay
We develop a model to perform a cost–benefit analysis of bus fare subsidies under financial constraints that preclude the purchase of additional buses. Our model considers users’ costs in the provision of bus services and the lack of road pricing to internalize urban transport externalities in a context where financial constraints severely limit the institutional ability to plan and design the bus system. Because of financial constraints, the bus system can hardly accommodate demand during peak times: buses travel overcrowded, passengers cannot board the first bus to arrive at the bus stop, and they cannot arrive at their destination at their desired time. Another salient aspect of our model is the inclusion of motorcycles as a second private transport mode. Motorcycles are typical in many urban agglomerations in the emerging world and engender many negative transport externalities. According to our results, fare subsidies provide social benefits in Metropolitan Asunción. During peak hours, a higher subsidy is justified as the reduction of the unpriced external costs of substitute modes compensates for the increased cost created by an additional bus passenger. In the off-peak, a higher subsidy is justified (i) as the higher frequencies induced by the new bus ridership reduce waiting times and (ii) because of the reduction of the unpriced external costs of substitute modes. Although our model does not explicitly include inequality aversion, we discuss the distributional aspects of subsidies in the context of middle-income countries.